refactor(api): split flow executor god class into per-node executor registry (ATT-428)
Extract each flow node type's logic into its own NodeExecutor strategy
under flows/node-executors, dispatched via a compile-time-exhaustive
registry built in the service constructor. Pure helpers (topicMatches,
heartbeatKey) and the NodeExecutor interface/context moved to shared
modules. The service keeps its DI signature and public surface and now
acts as a thin orchestrator (lifecycle, cron, event handlers, flow
traversal). Behavior preserved; existing tests stay green.
